+++ source/mysql/connection.d
@@ -97,6 +97,10 @@ package:
 
 	ulong _insertID;
 
+	// This gets incremented every time a command is issued or results are purged,
+	// so a ResultSequence can tell whether it's been invalidated.
+	ulong _lastCommandId;
+
 	// Whether there are rows, headers or bimary data waiting to be retreived.
 	// MySQL protocol doesn't permit performing any other action until all
 	// such data is read.
@@ -237,6 +241,8 @@ package:
 			"There are result set elements pending - purgeResult() required.");
 
 		scope(failure) kill();
+
+		_lastCommandId++;
 
 		if(!_socket.connected)
 		{
@@ -833,6 +839,8 @@ public:
 	{
 		scope(failure) kill();
 
+		_lastCommandId++;
+
 		ulong rows = 0;
 		if (_headersPending)
 		{
@@ -913,6 +921,10 @@ public:
 	/// After a command that inserted a row into a table with an auto-increment
 	/// ID column, this method allows you to retrieve the last insert ID.
 	@property ulong lastInsertID() pure const nothrow { return _insertID; }
+
+	/// This gets incremented every time a command is issued or results are purged,
+	/// so a ResultSequence can tell whether it's been invalidated.
+	@property ulong lastCommandId() pure const nothrow { return _lastCommandId; }
 
 	/// Gets whether rows are pending
 	@property bool rowsPending() pure const nothrow { return _rowsPending; }